util: migrate column-mapping and ddl-checker from tidb-tools to tidb

+++ util/column-mapping/column.go
@@ -1,4 +1,4 @@
-// Copyright 2018 PingCAP, Inc.
+// Copyright 2022 PingCAP, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -8,6 +8,7 @@
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
@@ -478,7 +479,7 @@ func partitionID(info *mappingInfo, vals []interface{}) ([]interface{}, error) {
 		return nil, errors.NotValidf("id must less than %d, greater than or equal to 0, but got %d, which is", maxOriginID, originID)
 	}
 
-	originID = int64(info.instanceID | info.schemaID | info.tableID | originID)
+	originID = info.instanceID | info.schemaID | info.tableID | originID
 	if isChars {
 		vals[info.targetPosition] = strconv.FormatInt(originID, 10)
 	} else {
